+++ pyembed/core/discovery.py
@@ -1,4 +1,26 @@
-from rembed.core import REmbedError
+# The MIT License(MIT)
+
+# Copyright (c) 2013-2014 Matt Thomson
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+from pyembed.core import PyEmbedError
 
 import requests
 from bs4 import BeautifulSoup
@@ -17,7 +39,7 @@ MEDIA_TYPES = {
 FORMATS = {MEDIA_TYPES[format]: format for format in MEDIA_TYPES}
 
 
-class REmbedDiscoveryError(REmbedError):
+class PyEmbedDiscoveryError(PyEmbedError):
 
     """Thrown if there is an error discovering an OEmbed URL."""
 
@@ -33,21 +55,21 @@ def get_oembed_url(url, format=None, max_width=None, max_height=None):
     :param max_height: (optional) the maximum height of the embedded resource.
 
     :returns: OEmbed URL for the resource.
-    :raises REmbedDiscoveryError: if there is an error getting the OEmbed URL.
+    :raises PyEmbedDiscoveryError: if there is an error getting the OEmbed URL.
     """
     media_type = __get_type(format)
 
     response = requests.get(url)
 
     if not response.ok:
-        raise REmbedDiscoveryError(
+        raise PyEmbedDiscoveryError(
             'Failed to get %s (status code %s)' % (url, response.status_code))
 
     soup = BeautifulSoup(response.text)
     link = soup.find('link', type=media_type, href=True)
 
     if not link:
-        raise REmbedDiscoveryError('Could not find OEmbed URL for %s' % url)
+        raise PyEmbedDiscoveryError('Could not find OEmbed URL for %s' % url)
 
     discovered_url = __format_url(link['href'], max_width, max_height)
 
@@ -60,7 +82,7 @@ def __get_type(format):
     elif format in MEDIA_TYPES:
         return MEDIA_TYPES[format]
 
-    raise REmbedDiscoveryError(
+    raise PyEmbedDiscoveryError(
         'Invalid format %s specified (must be json or xml)' % format)