+++ pyembed/core/test/consumer_test.py
@@ -1,4 +1,26 @@
-from rembed.core import consumer
+# The MIT License(MIT)
+
+# Copyright (c) 2013-2014 Matt Thomson
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+from pyembed.core import consumer
 
 from hamcrest import assert_that, equal_to
 from mock import patch, Mock
@@ -7,8 +29,8 @@ import pytest
 
 
 def test_should_discover_and_get_oembed_url():
-    with patch('rembed.core.discovery.get_oembed_url') as mock_get_url, \
-        patch('rembed.core.parse.parse_oembed') as mock_parse, \
+    with patch('pyembed.core.discovery.get_oembed_url') as mock_get_url, \
+        patch('pyembed.core.parse.parse_oembed') as mock_parse, \
             patch('requests.get') as mock_get:
 
         mock_get_url.return_value = (
@@ -32,8 +54,8 @@ def test_should_discover_and_get_oembed_url():
 
 
 def test_should_discover_and_get_oembed_url_with_max_width_and_height():
-    with patch('rembed.core.discovery.get_oembed_url') as mock_get_url, \
-        patch('rembed.core.parse.parse_oembed') as mock_parse, \
+    with patch('pyembed.core.discovery.get_oembed_url') as mock_get_url, \
+        patch('pyembed.core.parse.parse_oembed') as mock_parse, \
             patch('requests.get') as mock_get:
 
         mock_get_url.return_value = (
@@ -59,9 +81,9 @@ def test_should_discover_and_get_oembed_url_with_max_width_and_height():
 
 
 def test_should_raise_error_on_request_error():
-    with patch('rembed.core.discovery.get_oembed_url') as mock_get_url, \
+    with patch('pyembed.core.discovery.get_oembed_url') as mock_get_url, \
         patch('requests.get') as mock_get, \
-            pytest.raises(consumer.REmbedConsumerError):
+            pytest.raises(consumer.PyEmbedConsumerError):
 
         mock_get_url.return_value = (
             'json', 'http://example.com/oembed?format=json')
@@ -75,8 +97,8 @@ def test_should_raise_error_on_request_error():
 
 
 def test_should_embed():
-    with patch('rembed.core.consumer.get_oembed_response') as mock_get, \
-            patch('rembed.core.render.render_response') as mock_render:
+    with patch('pyembed.core.consumer.get_oembed_response') as mock_get, \
+            patch('pyembed.core.render.render_response') as mock_render:
         response = Mock()
         response.title = 'hello'
 
@@ -91,8 +113,8 @@ def test_should_embed():
 
 
 def test_should_embed_with_max_width_and_height():
-    with patch('rembed.core.consumer.get_oembed_response') as mock_get, \
-            patch('rembed.core.render.render_response') as mock_render:
+    with patch('pyembed.core.consumer.get_oembed_response') as mock_get, \
+            patch('pyembed.core.render.render_response') as mock_render:
         response = Mock()
         response.title = 'hello'
 
@@ -107,8 +129,8 @@ def test_should_embed_with_max_width_and_height():
 
 
 def test_should_embed_with_custom_template_dir():
-    with patch('rembed.core.consumer.get_oembed_response') as mock_get, \
-            patch('rembed.core.render.render_response') as mock_render:
+    with patch('pyembed.core.consumer.get_oembed_response') as mock_get, \
+            patch('pyembed.core.render.render_response') as mock_render:
         response = Mock()
         response.title = 'hello'