+++ pyembed/core/test/parse_test.py
@@ -1,4 +1,26 @@
-from rembed.core import parse
+# The MIT License(MIT)
+
+# Copyright (c) 2013-2014 Matt Thomson
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+from pyembed.core import parse
 
 from hamcrest import assert_that, contains_string, equal_to, none
 import pytest
@@ -102,10 +124,10 @@ def test_should_get_none_for_missing_element(format, function):
 
 
 def test_should_raise_error_for_unknown_type(format, function):
-    with pytest.raises(parse.REmbedParseError):
+    with pytest.raises(parse.PyEmbedParseError):
         get_response(format, function, 'unknown')
 
 
 def get_response(format, function, fixture='link'):
-    filename = 'rembed/core/test/fixtures/parse/%s.%s' % (fixture, format)
+    filename = 'pyembed/core/test/fixtures/parse/%s.%s' % (fixture, format)
     return function(open(filename).read())