+++ pyembed/core/test/response_test.py
@@ -1,0 +1,45 @@
+# The MIT License(MIT)
+
+# Copyright (c) 2013-2014 Matt Thomson
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+from pyembed.core import response
+
+from hamcrest import assert_that, equal_to
+import pytest
+
+
+def test_should_load_from_dictionary():
+    values = {'type': 'link', 'version': '1.0'}
+    oembed_response = response.OEmbedResponse(create_value_function(values))
+
+    assert_that(oembed_response.type, equal_to('link'))
+
+
+def test_response_should_be_immutable():
+    values = {'type': 'link', 'version': '1.0'}
+    oembed_response = response.OEmbedResponse(create_value_function(values))
+
+    with pytest.raises(TypeError):
+        oembed_response.type = 'photo'
+
+
+def create_value_function(values):
+    return lambda field: values[field] if field in values else None