+++ rembed/core/__init__.py
@@ -1,3 +1,0 @@
-class REmbedError(Exception):
-
-    '''Generic error class for REmbed.'''